Remove helium_logging.h and just use syslog.
syslog is pretty portable so I don't even know why I felt this was
necessary in the first place. Clients that really want to read libhelium
debug logs can call openlog() themselves.
