+++ samples/shell.c
@@ -1,6 +1,8 @@
 #include <stdio.h>
 #include <string.h>
 #include <inttypes.h>
+#include <syslog.h>
+
 
 #include <openssl/evp.h>
 #include <openssl/bio.h>
@@ -10,13 +12,15 @@
 
 void test_callback(const helium_connection_t *conn, uint64_t sender_mac, char * const message, size_t count)
 {
-  helium_dbg("Function-pointer callback got %s %zd\n", message, count);
-  helium_dbg("Mac address is %" PRIu64 "\n", sender_mac);
+  syslog(LOG_USER, "Function-pointer callback got %s %zd\n", message, count);
+  syslog(LOG_USER, "Mac address is %" PRIu64 "\n", sender_mac);
 }
 
 int main(int argc, char *argv[])
 {
-  helium_logging_start();
+  openlog("libhelium", LOG_PERROR | LOG_NDELAY | LOG_PID, LOG_USER);
+  atexit(closelog);
+  
   char *proxy = NULL;
   //helium_token_t token = "abcdefghijklmnop";
   helium_connection_t *conn = helium_alloc(NULL);
@@ -30,7 +34,7 @@ int main(int argc, char *argv[])
 
 #if HAVE_BLOCKS
   helium_open_b(conn, proxy, ^(const helium_connection_t *conn, uint64_t mac, char *msg, size_t n) {
-      helium_dbg("Block callback got %zu bytes from message %s", n, msg);
+      syslog(LOG_USER, "Block callback got %zu bytes from message %s", n, msg);
   });
 #else
   helium_open(conn, proxy, test_callback);
@@ -47,10 +51,10 @@ int main(int argc, char *argv[])
       helium_base64_token_decode(token_in, strlen((char*)token_in), token);
       if (strncmp("s", message, 1) == 0) {
         int  err = helium_subscribe(conn, mac, token);
-        helium_dbg("subscribe result %d\n", err);
+        syslog(LOG_USER, "subscribe result %d\n", err);
       } else {
         int  err = helium_send(conn, mac, token, (unsigned char*)message, strlen(message));
-        helium_dbg("send result %d\n", err);
+        syslog(LOG_USER, "send result %d\n", err);
       }
     } else {
       // invalid line, consume it
